Refactor: 홈 & 캘린더 페이지 디자인 및 API 수정 (#17)
* feat: 홈, 아카이브 페이지 구현

* feat: 주제, 상세 페이지 기초작업&디자인 구현

* feat: 주제,상세페이지 구현& 관련 팝업 구현 (수정 팝업 디자인 x)

* Style: 코드 정렬

* Feat: 캘린더 페이지 구현

* Feat: 검색 기능 구현 & Style: 홈,아카이브 디자인 수정

* Refactor: 수정 팝업 일정 입력 구조 변경

* Refactor: 홈 페이지 버튼 navigate  수정

* Feat: 홈& 아카이브 api연동(목업포함)

* Feat: 상세& 그룹 페이지 api연동 + 삭제 팝업 api연동 (목업 포함)

* Feat: 캘린더 api 연동중

* Feat: 캘린더& 수정 팝업 api연동 완료 (목업 포함) + Style: 피드백 디자인 수정

---------

+++ src/api/groupApi.js
@@ -3,6 +3,7 @@ import { mockGroupDetail } from './mockGroupDetail';
 
 const USE_MOCK = true;
 
+// 그룹 상세 조회
 export const getGroupDetail = async (groupId, page = 0) => {
   if (USE_MOCK) {
     return mockGroupDetail;
@@ -27,6 +28,53 @@ export const deleteGroup = async (groupId) => {
     };
   }
 
+  const response = await api.delete(`/api/group/${groupId}`);
+
+  return response.data;
+};
+
+// 그룹 수정
+export const updateGroup = async (groupId, groupName, groupCategory) => {
+  if (USE_MOCK) {
+    console.log('[MOCK] 그룹 수정 요청:', {
+      groupId,
+      groupName,
+      groupCategory,
+    });
+
+    return {
+      isSuccess: true,
+      code: 'COMMON_200',
+      message: '요청에 성공했습니다.',
+      result: {
+        groupName,
+        groupCategory,
+      },
+    };
+  }
+
+  const response = await api.patch(`/api/group/${groupId}`, {
+    groupName,
+    groupCategory,
+  });
+
+  return response.data;
+};
+
+// 그룹 카테고리 목록 조회
+// TODO: 팀원 API 파일 받으면 이 함수는 삭제하고 팀원 함수 import
+export const getCategoryList = async () => {
+  // 카테고리 목록 임시 목업
+  if (USE_MOCK) {
+    return {
+      isSuccess: true,
+      code: 'COMMON_200',
+      message: '요청에 성공했습니다.',
+      result: ['STUDY', 'SCHOOL', 'DAILY', 'ETC'],
+    };
+  }
+
+  const response = await api.get('/api/group/categories');
   // 실제 백엔드 연동
   const response = await api.delete(`/group/${groupId}`);
